superU_assignment_1/requirements.txt
asgiref==3.7.2
cffi==1.15.1
cryptography==41.0.1
Django==4.2.2
django-cors-headers==4.1.0
djangorestframework==3.14.0
jwt==1.3.1
pycparser==2.21
PyJWT==2.7.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
